Update web app OAuth example
Closes #18

The example is still ugly. Needs a rewrite. But it at least makes
clearer, now, how request tokens/secrets need to be handled.

Since we don't want to set up a database for the example, we store
access token/secret in a cookie. That's really bad practice. Noted in
the example.